Add missing documentation for state options for "image set"
The current "image set" with the state option behaves as follows: if you
use --project, it updates the membership status of the given project.

When this command was first added, it had the following description:

  If --project is passed, update the membership status for the given project

However, the description was missed when moving the documentation for
which command to the parser for each command method in commit f055fe67c.
Correct this oversight.

Co-authored-by: JAE YONG LEE <jaeljy135@gmail.com>
Change-Id: I18a29a19ce973971f68a52cbf4020cfa324c7a35