Migrate osc-tox-py3N-tips to Python 3.8
python-novaclient 18.0.0 removed support for Python 3.6 and 3.7 and
can't be installed in Python 3.6.

This change migrates the osc-tox-py3N-tips from Python 3.6 to Python
3.8 as Python 3.8 is the minimum version currently supported.

This also adds Python 3.9 to supported runtime list because it has been
tested since stable/yoga.

NOTE:
The osc-tox-py36-tips job is used by cliff, so the subsequent patch
for the cliff repo should be merged timely.

Change-Id: Id7a8f3ca5b704102aa18eb6bf98ae9ada005d013