Do not sort subnet dns_nameservers field
When using table output format, the dns_nameservers field
of a subnet is sorted, but it should not be as the order
is important.

Created an UnsortedListColumn() class in subnet.py so the
output is correct.

Updated the unit test accordingly to verify the order is
correct when an entry is removed.

Change-Id: I60a15a944f83549738305dd025db38ff8e165be7
Closes-bug: #2053201