Rename openstack volume delete --purge -> --cascade
This flag is called "cascade" in the Cinder API.
The flag "purge" doesn't really communicate an obvious
meaning in the context of volume deletion. It also
has the danger of implying some kind of behavior about
volume wiping that does not exist.

Rename this flag to "--cascade" and preserve "--purge"
as a hidden flag for compatibility.

Change-Id: I8de27811222c17155697073fb9c512746d009266
Signed-off-by: Eric Harney <eharney@redhat.com>
Co-authored-by: Stephen Finucane <stephenfin@redhat.com>