volume: Remove Cinder v1 support
The Cinder v1 API was removed in Queens [1]. Its replacement, the v2
API, has existed since Grizzly [2]. More importantly, the v1 commands
are implemented using python-cinderclient but support for the v1 API was
removed from python-cinderclient in Train [3], meaning none of these
have worked since then. Clearly if no one has noticed or cared in the 6
years or so since that happened, it's safe to say we can delete these
commands.

[1] https://opendev.org/openstack/cinder/commit/3e91de956e1947a7014709010b99df380242ac74
[2] https://opendev.org/openstack/cinder/commit/75ca60f619c953df1d95ff2eab7de78f7d5aebe8
[3] https://opendev.org/openstack/python-cinderclient/commit/2189e5702b7ba91a87e1db21024799e1520d8ad0

Change-Id: Ibe1cd6461d2cb78826467078aa17272f171746aa
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>